Seeder

+++ CthulhuWizard/CthulhuWizard.Application/Requests/Equipments/Queries/GetEquipments/GetEquipmentsQuery.cs
@@ -4,5 +4,5 @@ namespace CthulhuWizard.Application.Requests.Equipments.Queries.GetEquipments;
 
 public class GetEquipmentsQuery : IQuery<List<EquipmentDto>> {
 	public string? Name { get; set; }
-	public int? Price { get; set; } = null;
+	public double? Price { get; set; } = null;
 }
